+++ tasks/common/Cargo.toml
@@ -16,8 +16,6 @@ doctest = false
 console = { workspace = true }
 
 project-root = { workspace = true }
-serde        = { workspace = true, features = ["derive"] }
-serde_json   = { workspace = true }
 similar      = { workspace = true }
 
 ureq = { workspace = true, features = ["tls", "json"] }
