+++ tasks/transform_conformance/Cargo.toml
@@ -32,5 +32,4 @@ oxc_diagnostics  = { workspace = true }
 walkdir    = { workspace = true }
 pico-args  = { workspace = true }
 indexmap   = { workspace = true }
-serde      = { workspace = true }
 serde_json = { workspace = true }
